drivers/media/platform/exynos/mfc: Avoid shift overflow
./drivers/media/platform/exynos/mfc/s5p_mfc_enc_internal.h:1606:17: warning: signed shift result (0x100000000) requires 34 bits to represent, but 'int' only has 32 bits [-Wshift-overflow]
                .maximum = (2 << 31) - 1,
                            ~ ^  ~~

.maximum is a uint_64t, so this patch addresses the issue by casting 2 to a unint_64t before
the shift is done.

+++ drivers/media/platform/exynos/mfc/s5p_mfc_enc_internal.h
@@ -1603,7 +1603,7 @@ static struct v4l2_queryctrl controls[] = {
 		.type = V4L2_CTRL_TYPE_INTEGER,
 		.name = "Get extra buffer size",
 		.minimum = 0,
-		.maximum = (2 << 31) - 1,
+		.maximum = ((uint64_t)2 << 31) - 1,
 		.step = 1,
 		.default_value = 0,
 	},
